REF: Split GROMACS helpers into import and export
