+++ setup.cfg
@@ -18,6 +18,7 @@ per-file-ignores =
     openff/interchange/smirnoff/_*:F821
     openff/interchange/components/*:F821
     openff/interchange/__init__.py:F401
+    openff/interchange/interop/gromacs/__init__.py:F401
     openff/interchange/tests/data/*:INP001
     plugins/*:INP001
